CNN Update
Changed code to use the US video URL so it will not try to pull up the
International main video page. Both return basically the same thing, but
International has a section with an error in the code.

Also added code to prevent an error if encounters a bad section like in
the International version of videos in the future.